Move EventQueue into its own module, fix bug when tessellating several paths with the same context.
